multiversion: Remove use of the $globals temporary JSON cache file
This cache was introduced many years ago as PHP didn't have a (good)
opcache, and the logic to iterate these large arrays was costly
and expensive, both due to the cost of the essential operations in
PHP as well as due to the (slower) implementation of SiteConfiguration
we had at the time.

By now, the time we take to check the various mtimes of source files,
and open/read/parse the JSON file, and allocate it recursively into
memory on every request, is longer than it would to simply read
the source files (given php-opcache) and iterate them accordingly.
and return the values directly.

Bug: T169821
Bug: T311788
Change-Id: Ic0dbcba9f60f20a172cadbf0d821eb665ef9136c