multiversion: Move labs-overrides responsibility to getStaticConfig()
This means getConfigGlobals() will no longer have to mutate the
given SiteConfiguration object, and the callers in CS.php and
wiki.php will become a bit easier to read and reason about, e.g.
a series of assignments and then a call to extract the globals.

In the next changes:

* Move $conf->settings assignment to CS.php.
* Move callers to getConfigGlobals().
* Flatten getMWConfigForCacheing() into getConfigGlobals().

Bug: T308932
Change-Id: I6e97d39a3931073c65af58e2ff523217dbd8f467