CirrusTest: Remove reference to 'unittest' realm
This isn't a registered realm. It was not used anywhere, except for
one condition where it is compared against "!= labs" and thus
effectively the same as production. Use that directly instead.

This is in prep for I6e97d39a3931073c which cleans up some
settings-loading code, and has as side-effect that some code which
is currently in CommonSettings.php would become re-usable and move
to MWConfigCacheGenerator, and this code uses $realm to load e.g.
'labs' or 'traindev' settings files, and thus expects such files
to exist, and thus CI fails there unless we fix this first.

Bug: T308932
Change-Id: Ib7e87a5c97d8feffc9bc36b2ab59089dde6c24b6