Remove legacy wgRC2UDPPrefix overrides for private wikis
Follows-up b2aae61de (I249fa712a), in which we locked legacy site/lang
values in-place, with the rest based to use canonical servername.

What I missed there is that many of these, do not have wmgUseRC2UDP
enabled and so the value of wgRC2UDPPrefix is not actually useful
to keep unchanged as it isn't used on these private wikis.

Change-Id: I0da9c662693ee96f708401fb76f18317b7149e11