Pin wgCheckUserLogReasonMigrationStage to read and write old
Pin the CU log reason migration stage to read and write old because the
columns cul_reason_id and cul_reason_plaintext_id do not exist on
WMF wikis. The cul_reason field is being migrated to use the
comment table and a patch depends on this because it starts to
write new on wikis with the field in the DB. Instead of checking
for the existence of the field, that may be different depending on
the exact server used, a config is used to ensure writing only occurs
once the field definitely exists.

Bug: T233004
Change-Id: I6d3a50799808eb473d37988eedbf33a8d1dddbcd