Explicitly set wgMessageCacheType=mcrouter (avoid newAnything in prod)
The MessageCache service wiring in core, is currently the only place
that I could find in flame graphs where we end up calling
ObjectCache::newAnything, since $wgMessageCacheType = CACHE_ANYTHING
by default and we don't override it in production.

Make this more explicit to remove any doubt where this goes in prod,
e.g. it's not going to be written to wgParserCacheType which is
included in the anything-chain.

Bug: T186673
Change-Id: Id9707db2273b31e12c5ed9ae22c852176cad7c8f