cirrus: Handle transition to elasticsearch 7.10
The integration between mediawiki and elasticsearch is only compatible
with either elastic 6.8 or 7.10, but not both at the same time. The
changeover will ride the train on 1.39.0-wmf.28. Expected state
is that when wmf.28 starts rolling out the codfw cluster will be
running 7.10 and the eqiad cluster will be running 6.8. The set of
servers to contact needs to change as the branch changes.

Additionally handles ensuring write compatability from both branches
to both clusters running different version of elasticsearch. The
elastic writes from wmf.27 are compatible with both 6.8 and 7.10, but
the writes from wmf.28 need a special compatability transport enabled
to ensure the writes work against 6.8.

Cloudelastic does not need any special handling here. It will be
upgraded in entirety prior to the wmf.28 branch cut and the pre-wmf.28
branches are fully compatible for writes with elastic 7.10.  No
querying against cloudelastic is performed from prod, and as such
no compatibility is required there.

Change-Id: Id26558ca2c43dcf550c294824674365929636184