testwiki: Add mediawiki.edit_attempt stream
This stream is published to by all client-side EditAttemptStep
instruments. More generally, the stream is published to by all
instruments dispatching "eas." events via the Metrics Platform [0].

As we did with the introduction of the mediawiki.web_ui.interactions
stream in Ie06e57eb477, define the mediawiki.edit_attempt stream for all
wikis with a sampling rate of 0 and setting the sampling rate to 1 on
testwiki and all Wikipedias on the Beta Cluster.

[0] https://codesearch.wmcloud.org/search/?q=%27eas%5C.&files=%5C.js%24

Bug: T309013
Change-Id: I070d826f63dae9e882137fd3d9bb3a76f6622a50