Remove wmgEntityUsageModifierLimitsStatement on cebwiki
We added this limit before enabling statement usage tracking on cebwiki,
to lower the theoretical upper limit for wbc_entity_usage rows as a
result of that change. However, in actual fact only very few pages reach
this lower limit (139 'C' usages as of 2022-09-28 14:29:25 UTC), so we
can safely remove it without increasing the number of wbc_entity_usage
rows too much.

Bug: T296384
Change-Id: I829a0f265b5f9190d50f2f0134747c693f0076d2