cirrus: Don't configure cloud clusters for private wikis
Other protections (wgCirrusSearchWriteClusters) ensure these clusters
are not written to, but it would be safest if they weren't even
configured in the first place.

This reverts commit f84d4ff38f6c58bf9a66d8389d729439804d5093.

Reason for revert:  Scap
deploy ran php7.2, which doesn't support trailing commas, although
the rest of prod is 7.4. Prior patch updated to emove the trailing
comma.

Change-Id: I4cb6b61f3bf4a94e381763319983c8600292ada9