cirrus: remove cross-dc poolcounter increases
With read traffic split between multiple datacenters we no longer
have a single poolcounter limiting read traffic to a single
elasticsearch cluster. In instances where we send all traffic to a
single cluster there is likely to be both local and remote traffic
going to that cluster.

As such, stop increasing pool counter limits when talking to a remote
cluster. That cluster will already be under some load. This may result
in some poolcounter based rejection the next time we do a test
involving sending all traffic to a single DC, values can be retuned
as necessary then.

Change-Id: I4e193574b7157a69047c65b8a9981f258df492c8