cirrus: Drop client side connect timeout config
This was set to 5 seconds for local clusters and 10 seconds
for remote. Nothing about cross-dc operations should require
more than 5 seconds to make a connection. Stop setting this
value and accept the extension default value, which is also
5 seconds.

This is prep work for sending all cirrus read traffic through
a single dns-discovery based cluster. In that case we wont be
able to set per-cluster values as it will be determined
externally.

Bug: T143553
Change-Id: Icecab2700d8c2ae44dfab6b6ea7d33cbdba47309