Beta: Enable parsoid cache warming.
This causes a parsoid rendering of the page to be saved to the parser
cache on every edit, in addition to the output of the old parser.

NOTE: this doubles the amount of data written to the parser cache.
On the beta cluster, the parser cache is memcached only, no database.

Bug: T320535
Change-Id: I0ef25393fd3df9a0ff9e774482f293ab471f2e94