Use OpenSSL for PBKDF2 password hashing
Now that 1.40.0-wmf.5 is active on all sites, use the new class added in
core commit 47241a3520d5b4bf, which is more efficient.

The format of type "pbkdf2" password hashes does not change, and the new
class (Pbkdf2PasswordUsingOpenSSL) is fully compatible with the existing
hashes stored in Wikimedia databases, all of which specify "sha256" or
"sha512" as the hash algorithm. Increasing the strength parameters, if
desired (see T234987), would be done in a separate commit.

Bug: T320929
Change-Id: I4af95a81e5ae5eaac3ad73054e3c07cca51e70c3