cirrus: Correct comments in ProductionServices.php
These comments say that some resources are not pointed where they
are expected to point, but on review (see ticket) everything lines up
in the ways that it should. Remove the misleading comments.

While here, re-order the clusters so the ports are in numerical order,
it's easier and less surprising to read this way.

Bug: T262630
Change-Id: Ibbfd9bcc1542fa6f7aa31d9f5b5cca54bd8f8a8b