Enable parsoid cache warming on testwiki.
This will cause the parser cache entry for parsoid output to be
populated on every edit. This will allow VisualEditor to quickly get the
current rendering of the page.

NOTE: this will (eventually) roughly douple the amount of space needed
by the parser cache. Before enabling this on large wikis we need to be
sure that we have enough capacity available and keep DBAs posted.

Bug: T320535
Change-Id: I8a68c931ace5874d1e65020fc74df762fbebdd1f