Set VisualEditorDefaultParsoidClient for dewiki-beta mad testwiki
We want to make the VE API talk to Parsoid directly instead of calling
out to RESTbase. Per I3b73431fcca we now have a config setting to
control this directly. We should try it out on beta and an testwiki.

Note that we also want to keep most of beta using the old mode which is
still in use on the live wikis. We try the new mode on dewiki beta only.

Bug: T320531
Change-Id: I1d825d2927857a88e555c994508b428b0ee00920