Wikidata: don't show Vector search thumbnails
wbsearchentities does not yet contain page thumbnail data, so showing
the placeholder for every result is a bit pointless.

Both settings are beings set, because setting only one would mean that
the default value of the other setting gets dropped. But using wiki-key
that is prefaced by `+` doesn't work either, because of T324859.

Bug: T316093
Change-Id: Ic28029dc219b62709b91facb7fb0e7fd17ed2622