Temporary identity mappings for title case ligatures
Both the source and destination of a move needs to be valid for the
script to work. For example we have enwiki Ǉubišǌa → ǈubišǌa, the source
being a broken artifact of a previous migration which will become
invalid. By mapping ǈ to itself, the page can be moved prior to the
completion of the migration.

Bug: T292552
Change-Id: I7a4cbaf26245e5bf123f33cc7b169e9dc1f2d6a5