testwiki: Add mediawiki.visual_editor_feature_use stream
This stream is published to by all client-side VisualEditorFeatureUse
instruments. More generally, the stream is published to by all
instruments dispatching "vefu." events via the Metrics Platform [0].

Similar to adding mediawiki.edit_attempt stream in I070d826f63d,
the mediawiki.visual_editor_feature_use stream for all wikis have a
sampling rate of 0 and we set the sampling rate to 1 on testwiki and
all Wikipedias on the Beta Cluster.

[0] https://codesearch.wmcloud.org/search/?q=%27vefu%5C.&i=nope&files=%5C.js%24&excludeFiles=&repos=

Bug: T309602
Change-Id: I47eaccfac86500052ab0a519543a988e57421c86