[Growth] Do not override wgGEMentorshipUseIsActiveFlag
GrowthExperiments' extension.json file sets
wgGEMentorshipUseIsActiveFlag to true by default.

Since the relevant table column is now in production,
there is no need to override the flag to false
in production.

Removing the option instead of changing to true,
as the config option will soon go away
as part of housekeeping (T319037).

Bug: T318457
Change-Id: Ic14edd5277ebc2e416cfb264fa4c829e266a6664