Set parser cache write propability for /page/html endpoint.
Control the probability with which Parsoid's /{domain}/v3/page/html endpoint
writes to the parser cache. This endpoint is hit by RESTbase when pre-populating
its cache after a page is edited. Setting this to 1 cause all edits to trigger
a parser cache write, putting full pressure on the parser cache.
This needs to be enabled gradually.

Needed-By: Ie430acd0753880d88370bb9f22bb40a0f9ded917
Bug: T322672.
Change-Id: I4c1c82f4bae59d786fed92b743ee7a7216ed24b1