cirrus: Enable document size limiting
CirrusSearch can generate search documents that are unbounded in size.
Previously we limited text from file sources only on commonswiki, now
we are expanding that to a general document size limit applied to all
wikis. This should bring consistency and stronger expectation
management around what goes into the search engine.

With the new document size limiter profile it is no longer necessary
to set a direct limit for file_text.

Bug: T323687
Change-Id: Id0913b0acd06ab569fec15d3f5e237598401c660