hewiki: enable parser cache writes for parsoid's page/html endpoint.
Parsoid's page/html endpoint is used by RESTbase when it is notified
that a page changed, and the HTML needs to be updated. Since the
endpoint is hit for every edit, enabling parser cache writes on it
amounts to enabling parsoid cache warming.

This is done in preparation of switching VisualEditor to direct mode.
We start with hewiki because it has enough VE usage to provide
meaningful numbers, but is small enough that it won't cause problems in
case something is off with the parser cache.

Bug: T322672
Bug: T320534
Bug: T320529
Change-Id: Ic3a7ed5a8ff1dd877d2f8e7d057d5f095e49d3af