Disable writing parsoid html to PC on commons and wikidata.
We are currently writing parsoid output to the parser cache for 5% of
requests to Parsoid's page/html endpoint, which is hit by RESTbase when
pages change. For commons and wikidata, this is useless, and causes of
lot of pointless cache writes. We should be smarter about filtering
these out, but for now, just disable parser cache writes for this
endpoint on these wikis.

Change-Id: I375d6c61a894bc54119d0b12f03909c6dcc49e11