Profiler: Refactor to make suitable for CLI again
== What

* Skip Excimer sample profiling on CLI (limit to web reqs and jobs).

* Remove `tideways_xhprof_enable` call for CLI. This condition existed
  by mistake. We don't want to debug-profile all CLI invocations.
  Fortunately, it didn't actually do that yet as Profiler.php isn't
  (yet) loaded on CLI, but soon will be.

* Keep setting $wmgProfiler on CLI, but without 'running' => true.

* Remove support for "forceprofile=1" as GET query param,
  keep as XWD attribute.

== Why

Back in 2017, wmf-config/StartProfiler.php was refactored into
PhpAutoPrepend.php to address lack of visibility into early MW setup
and multiversion setup during web requests (T180183). This had the
side-effect of making the profiler unavailable to CLI invocations.

Then later in subsequent commits, the (unused) code for CLI kept getting
further simplified, including for "enable" to be called unconditionally
which doesn't make sense since nearly 100% of CLI invocations would not
be with the --profile option. Fortunately, since this file was never
loaded in this form in production for the CLI, it didn't actually
start the profile for all CLI maintenance scripts.

In even later commits, Excimer was added for sample profiling of
production traffic. This too didn't exclude CLI contexts so those would
be sampled too if we didn't disable that. We could certainly consider
enabling this at some point, but keep the status quo by disabling it
explicitly for now. One thing to think about if we enable it is whether
to include it in the excimer "all" entry point, and it might be useful
to prepend a virutal stack frame shared by all CLI scripts for easier
visualising and filtering.

While at it, remove legacy "forceprofile=1" query parameter.
This has been obsoleted by XWD forceprofile=1 attribute. Given that
it was already conditional on the header, there isn't a usability
benefit to enabling it via query parameter as by definition one would
already have to set the header (e.g. via curl) or via the browser
extension, both of which naturally offer the ability to set the
header attribute instead.

Bug: T253547
Change-Id: Ife7bded7480946ceed458ebfa78e523be544523e