wmf-config: add new revision-score streams for EventGate main
The ML team would like to replace the mediawiki.revision-score
stream with multiple/more-granular ones, to better segment
the model scores. The idea is to have separate streams for each
revscoring-based model, so that clients can subscribe to more
granular stream of scores. This new layout would also better fit
the Lift Wing architecture, in which we don't have a specific
entrypoint to score a single rev-id with multiple models.

Replacing mediawiki.revision-score would also mean moving ahead
in the direction of the ORES deprecation.

Bug: T317768
Change-Id: I7678f2e1df5a6aea225279993bd70009946b380d