Remove redundant definition of wgCheckUserEnableSpecialInvestigate
Remove redundant definition of wgCheckUserEnableSpecialInvestigate
in InitaliseSettings.php as the value specified is equal to the
default value of true in extension.json.

The default for this config was changed in c77285d which was merged
over a month ago (so the value is very unlikely to be changed back
to false).

The removal of this config value should make no change to wikis.
The testing method is to ensure that Special:Investigate is
accessible on any wiki that has CheckUser enabled (test is a
success if it is accessible).

Change-Id: Idb889d8c2982d7fbb049dc5e38a520f855f39819