multiversion: Create dblist-manage command for easy add/delete
The YAML files currently serve as only purpose to ease the addition,
removal and listing of dblists, at the cost of duplicating large
amounts of code in MWConfigCacheGenerator that are currently unused
as it was written for YAML-based management of InitialiseSettings.php.

Given the direction Platform Team is going in terms of using static
PHP files (after having considered YAML), and no current roadmap for
using this code, I'd like to simplify it for the time being and make
the current logic more debuggable and understandable to have clean
slate for future development.

The dblist data can be viewed live at
<https://noc.wikimedia.org/wiki.php>, or via the CLI by using
`dblist-manage list`.

The "add" command is agnostic about whether dblists are managed by
wiki or by dblist. It takes care of adding it to the right place.

CI validation is kept as-is.

The now-unused code will be removed in a follow-up commit for ease
of review.

Bug: T308932
Change-Id: I33135682558d94f4c17f9e7bc224d8a1045efb28