logos: Exclude logos/index.html from Git
The file was already outdated (dirty roundtrip) and takes about half
a minute to generate, which seems too slow to include in 'composer test'
for enforcement.

Given it is only used for local preview, exclude it so that it doesn't
have to be kept up-to-date in the repository.

Once generated locally, the file will survive other commits, updates,
and development so you won't have to regenerate it any more often
than today.

Change-Id: Ie3793f9f92ea9cd3968dd792f6b71dc7c1e7346f