speed-tests: Add captureSpeedtest.php script and publish 2023 snapshot
The last reference snapshot was from 2021. Create a new one snapshot
that uses Vector 22 and includes everything else that has changed
since 2021, which then makes it easy for experimental variants
to be minted from this one.

Also, automate the generation of these standalone static captures
to ease next time this needs to be done.

Change-Id: I8b95c96d8d808d8e5b8233fadc97eff2e19ec16a