Produce rc1.mediawik.page_change to eventgate-main
This will cause EventBus to produce these events to eventgate-main,
which in turns produces to kafka main clusters.  Messages
will be replicated to kafka jumbo, and then still in development
mediawiki-page-content-enrichment service (running in dse-k8s-eqiad)
can still use kafka jumbo for now.

This will help with other users that want to start working with the
page_change stream while we roll it out.

Change-Id: Ib5233feed2a6a292307afa884dc28e905752c0d7