tests: Improve diffConfig by sorting keys first
When settings are re-ordered in the same file, or e.g. moved to
another file, the diff was showing them as "changed" by removing the
line in one place and then re-adding it elsewhere. When this is just
a handful of lines for 1 wiki, that's easy to confirm, but when
it's thousands for every wiki/realm, it is infeasible to confirm
by eye.

Change-Id: I868329f3b9f543b99de0b49a211ae385d9319e15