Add static "Cleopatra" page to facilitate synthetic testing of 885362
Adds a "before" and "after" version of the "Cleopatra" page from
https://en.m.wikipedia.org/:

"before": Copy of the production page
"after": Adds a `srcset` attribute to each lazy loaded image in the "before" page

This will help determine the effect that 885362 has on metrics like HTML parse
time. For effects on payload size, see T293303#8630702.

Note: I'm planning for this to exist temporarily — I'm interested in the
synthetic test results and then will remove it. As a result, I didn't try to
make the CSS and JS static.

Bug: T326147
Bug: T293303
Change-Id: Idf8ac13644223a3caa65e6fc473ac2557d7578da