build: Change diffConfig to use git-stash instead of git-add
This allows both states to be be stored independently as commit drafts
relative to HEAD, which can then be compared as equals.

This avoids the problem where a deleted wiki or otherwise an entirely
removed or broken config file is wrongly reported as no diff.

== Previous logic ==

* Generate the "after" state on HEAD, and stage the output directory.
* Generate the "before" state on HEAD~1.
* git-diff the unstaged changes, reported with --reverse (-R) so that
  "after" really is after.

For the case where a new file is created in the "after" state, this
will correctly notice that the file was staged but appears deleted
in the "before" state as it won't be in the working directory at
the end when we compute the "before" state.

For the case where a file is deleted by the current commit in the
"after" state, the deleted file will not be staged as it doesn't
exist (and you can't stage the absence of a file). Then in the before
state, it will be created and remain as an untracked file in the
working directory. Untracked files, however, are not considered by
git-diff.

== New logic ==

* Generate the "after" state, stage it, and stash it.
  This effectively creates an commit-like index entry for it
  with the symbolic name `stash@{0}`, and resets the working
  directory back to HEAD.
* Generate the "before" state, stage it, and stash it also.
* git-diff between `stash@{0}` and `stash@{1}`.
  This will naturally consider each one fully, no relevant files
  are untracked as each one staged its own files.
  Also no need for --reverse.

Change-Id: Iebe656796782809337d5d639157728d2fc338b4d