filebackend: hotfix - make swift master follow the mediawiki master
Please note that while this is not strictly correct, it's better than
having this statically defined; also it follows what swift cloning does
anyways.

If we are right, this should mitigate the ill effects we've seen on
thumbnail reuploads.

The patch is simplistic to reduce risk when deploying it, as it's a
fix (we think) for an UBN! issue.

Bug: T330942
Change-Id: I90ede9787d432820c057be734955a51fd13d1f50