build: Fix missing git config for git-stash command
While having no commit message, stashes are commit-like in that they
have a (temporary) author and committer. This defaults to something
implied from the environment, but the default self hostnames in our
current Docker images are of a form that Git is uncomfortable using
by default. Set it to something generic (but valid) like 'localhost'
instead.

Assuming this works, this might be good to upstream to our
integation-config base images at some point.

Bug: T331020
Change-Id: Iff8c7e6b2a75a24d994655f550632921eb7b1407