Always write parsoid output to parser cache.
This has been at 50% for a week now, without problems. Let's try 100%.

Parsoid PC writes are at 6K per minute for the 50% setting,
so we expect 100% to translate to 12K/minute.
This is in addition to the 40k writes per minute to the main PC.

The next step will be to enable the cache pre-warming job. in WMF
production, that shouldn't lead to additional parsing, since RESTbase is
requesting parsoid renderings after every edit anyway, which (per this
patch) get written to the parser cache.

Bug: T320534
Change-Id: Id2c918aea259ccfb50619ca602ba5da269d6539d