build: add local typos check
I have introduced a typo checker back in 2016 (T133047) which is
implemented in the image `docker-registry.wikimedia.org/releng/typos`.
It all boils down to running:

  ! /usr/bin/git grep -E -I -f typos -- . ':(exclude)typos'

The leading exclamation mark (`!`) is to invert the exit code since `git
grep` exit code is 1 when it finds nothing which is expected when there
are no typos. To support Microsoft Windows user, wrap it with `sh`
(thanks Bartosz Dziewoński for the tip).

The container image is based on Debian Stretch which we should phase out
and it is a good opportunity to move the check directly in the source
repo.  The straightforward way is to add a "typos" composer script which
is called from "test" and will thus run on CI or locally. That saves
people to be caught by a CI error when they could have caught it
locally.

Once this is merged, we can remove the container image and the CI job
`operations-mw-config-typos-docker`.

Bug: T332121
Co-Authored-By: Bartosz Dziewoński <matma.rex@gmail.com>
Change-Id: Id4d837e62317b23c59b5b263574f2e543878b7c5