Remove possibly significant whitespace from robots.txt
While per RFC 9303 § 2.1, a group is terminated by a new user-agent
line or end of file, in practice not all implementations adhere to
the spec as well and may interpret the extra newline as a new group.

Since that group has no corresponding user-agent line, the disallow
lines don't actually do anything.

A notable buggy parser is the one shipped in the CPython standard
library, for example:

>>> import urllib.robotparser
>>> r = urllib.robotparser.RobotFileParser()
>>> r.set_url("https://en.wikipedia.org/robots.txt")
>>> r.read()
>>> r.can_fetch("*", "https://en.wikipedia.org/wiki/Main_Page")
True  # <-- correct
>>> r.can_fetch("*", "https://en.wikipedia.org/trap/foo")
False # <-- correct
>>> r.can_fetch("*", "https://en.wikipedia.org/wiki/Wikipedia:Checkuser/foo")
True  # <-- wrong
>>> r.can_fetch("*", "https://en.wikipedia.org/wiki/Wikipedia:Articles_for_deletion/Foo")
True  # <-- wrong

The simplest thing to do here is just remove the newline that was presumably
unintentionally added in e8123ad4fa.

Bug: T334038
Change-Id: Ibc1fc3491454a828ec88ce5dec2c359d7ea8e228