Expose the 'sfsblock-bypass' right so it can be assigned to global groups
StopForumSpam is installed in a limited set of projects at this
moment, Meta-Wiki not being one of them, so it is not possible
to grant this right to global groups.

This permission is currently required for stewards and global
sysops on projects that are in SFS 'enforce' mode and may happen
to be using a IP address identified as problematic by the SFS
extension. Since SFS blocking is pretty aggresive, it is blocking
as well administrator actions such as deletions, so this is needed
to unblock some global users' administrative work.

Bug: T334856
Change-Id: I2162a79d0953b450cf4c017951a46f5a257e6a3d