beta: Link to translations of CC BY-SA 4.0 where possible
Creative Commons makes translations of the license available under
/deed.LANGCODE . Use the $lang feature in the config system to link to
these pages for most wikis. CC doesn't have translations in all of our
languages, but that's fine because it falls back to English for
unrecognized language codes.

Don't use this $lang trick for special wikis, where $lang clearly is not
a language code (but is 'meta', 'commons', 'wikidata', etc)

Bug: T319064
Change-Id: I1be87853a1ccb0e7cffaa78c0b623caaaab6d581