Declare mediawiki.page_change.v1 stream and produce from Eventbus
This will declare the new stream, and cause EventBus to start producing to it.
Any existent downstream consumers will need to reconfigure to use the new stream.

This also changes the error stream name for the mediawiki-page-content-change-enrichment
job, so we'll need to change that configuration there too.

Once https://gerrit.wikimedia.org/r/c/mediawiki/extensions/EventBus/+/920377
is deployed, we can remove the wgEventBusStreamNamesMap override.

Bug: T336817
Bug: T326536
Change-Id: Ic19a8b369bbb7b12aa52d28cfff3e5360ae891db