ORES: add model versions configuration and thresholds
Add ORES models versions in config.
Due to the switch to use Lift Wing instead of ORES we add these in the configuration
since the ORES extension needs model versions during initialization phase in order to
populate the ores_models table.
ORES supports querying the models and balance specificity with sensitivity based on queries provided
in the configuration for each model. In order to facilitate the transition to Lift Wing we load these thresholds from configuration
instead of fetching them from ORES.

Bug: T319170
Change-Id: Ia770b3dc8b4ed2e6d32b75b1e898d9124997f520