Switch VisualEditor to not use RESTbase on small and medium wikis
We have been testing the new "direct" mode on hewiki and mediawiki.org
for several months now, it seems to work nicely. Cache warming jobs have
been enabled for small and medium wikis for a couple of weeks now, so
the parser cache should be prepared.

NOTE: enabling this will put pressure on the main stash, because it will
be used for storing content while editing. Based on the metrics gathered
from the equivalent cache in  RESTBase, we estimate a need about 140GB
of storage capacity for the VE edit stash once this is enabled for ALL
wikis. This is based on the measurement of 100 writes per second and a
24 hours TTL, with an average of 20KB per entry.

Bug: T320529
Change-Id: I3152fee1408bec5798c11b9a70e4cc0729444b15