Fix some mwscript bugs and clean up the style
The feature allowing mwscript to run a script with an absolute path was
broken, because wmfGetMWScriptWithArgs() was putting the path to the
script in $argv[0] and the basename of the same script in $argv[1],
despite executing run.php. When a relative path was given,
getMediaWikiCli() copied the path from $argv[0] to $argv[1], covering
up the error. When an absolute path was given, run.php tried to
execute the basename passed in $argv[1].

Here's how arguments are meant to work:

             |      |    $argv[0]          |     $argv[1]
-------------+------+----------------------+---------------------
Input        | php  | mwscript             | script.php
Old output   | php  | /path/to/script.php  |
New output   | php  | /path/to/run.php     | /path/to/script.php

So:

* Get rid of the basename thing
* Make the code easier to follow by reducing the number of points at
  which $argv is modified.
* Set the correct $argv prior to calling getMediaWikiCli().
* In the run.php case, have getMediaWikiCli() get the script name
  from $argv[1] instead of $argv[0].
* Also fix up $argv[0] in the $useOld case.
* Add --mwscript-debug-dump-argv feature so that $argv can easily be
  inspected in production.
* Rename some local variables.
* Use a consistent comment style.
* Include MWMultiVersion.php from the top of the file, like a normal
  person would.

Change-Id: I0ea50682345d1fec4939e82eb0fd1b3a5d538212