OAuthRateLimiter: Add rate limiting class for WME using LiftWing
This class is mainly intended to be used for WME's access to LiftWing
services, as they need more qps than the base `internal` class allows
for. Only specific tokens will be moved to this class for now.

There is additional rate-limiting config in the API GW itself, as well
as on the k8s side in LiftWing.

Bug: T338121
Change-Id: I837c165cad8f10f96d54214569cdca7893a4e580