Enable parser cache warming jobs for parsoid on enwiki
Enabling this for frwiki (I376e4aa20f9) went without incident, though we
saw a sharp rise in the enqueue rate for the parsoidCachePrewarm job, as
expected: the rate jumpted from barely noticable to close to 50 jobs per
second.

We want to enable this for all wikis soon, except probably for
wikidata and commons. Enabling it for enwiki is expeded to add roughly
200 jobs per second to the enqueue rate.

NOTE: When deplyoing this change, please monitor the "Job" section at the
bottom of the VE backend dashboard:
<https://grafana.wikimedia.org/d/OxxOv5K4k/ve-backend-dashboard>. Should
the wait time or concurrency fow too much, we will have to revert the
change and commit more hardware.

Bug: T329366
Change-Id: Iec57b50b7af48e135ff1895c78b3afb15fd0bf02