Remove old origin-with-crossorigin referrer policy
This was a misspelling in some old version of Safari between 2017 and
march 2018. These user agents are hardly in use any longer, but this
setting actually throws a console validation error in any modern
version of Safari and it wastes bytes.
It's no longer needed any longer in my opinion.

Bug: T338183
Change-Id: I66deda17f092f22e5b5916df1b18ed30f17697b8