Enable cache warming jobs for parsoid per default.
Enabling parsoid cache warming jobs everywhere except commonswiki and
wikidata. Enabling this for enwiki (Iec57b50b7af48) went without incident,
with less of an inccrease in the enqueue rate for the parsoidCachePrewarm
job than we expected (from 50 to 120). Job runner concurrency went from 4
to 9.

With this patch, we expect the enqueue rate to roughly double.

NOTE: When deplyoing this change, please monitor the "Job" section at the
bottom of the VE backend dashboard:
<https://grafana.wikimedia.org/d/OxxOv5K4k/ve-backend-dashboard>. Should
the wait time or concurrency fow too much, we will have to revert the
change and commit more hardware.

Bug: T329366
Change-Id: I607ae0f9be3518e3930bed617df86ceb35131ac3