Switch VisualEditor to not use RESTbase on English Wikipedia.
We have been testing the new "direct" mode on small and medium wikis
for several weeks now, it seems to work nicely. Cache warming jobs have
been enabled for all wikis now, so the parser cache should be prepared.

We enable this for enwiki to see what impact it will have on stash
usage. If all goes well, the plan is to enable it on all wikis soon.

NOTE: enabling this will put pressure on the main stash, because it will
be used for storing content while editing. Based on the metrics gathered
from the equivalent cache in  RESTBase, we estimate a need about 140GB
of storage capacity for the VE edit stash once this is enabled for ALL
wikis. This is based on the measurement of 100 writes per second and a
24 hours TTL, with an average of 20KB per entry.

Bug: T320529
Change-Id: I332f05177be421936d89a9bd10362ff889fe0515