cirrus: Enable analysis chain deduplication for wikibase
We've been having issues with WikibaseCirrusSearch indices for some
time. They are slow to create, can trigger master_timeout issues, and
generally don't behave like the other indices. We've tracked this down
to an issue in elasticsearch with how it handles large amounts of
index settings. By dramatically reducing the number of settings
through deduplication we can get the behaviour of these WBCS indices
mostly inline with the normal indices.

Bug: T334194
Change-Id: I2f5a3106cb4c90cd30bee48c647f8c0da85a4ff1