Parsoid: Disable PC writes on frwiki
The parsoid endpoints are primarily hit by restbase when pre-generating
page content. Having them write to the parser cache is redundant to
having ParsoidCachePrewarmJob do the same. If the parsoid endpoints no
longer do it, the jobrunners will have to do the parsing work.

This patch is an experiment to see the impact this would have on the
load on the jobrunner cluster. If it goes well we will try doing the
same on dewiki, enwiki and eventually for all wikis.

Bug: T339867
Change-Id: I8820a83d12c296ee5cecbee031452f7a312540ad