Parsoid: Disable PC writes on dewiki
The parsoid endpoints are primarily hit by restbase when pre-generating
page content. Having them write to the parser cache is redundant to
having ParsoidCachePrewarmJob do the same. If the parsoid endpoints no
longer do it, the jobrunners will have to do the parsing work.

This patch is an experiment to see the impact this would have on the
load on the jobrunner cluster. We need tio try it with somethign big
enough that the impact would be visible in the overall stats.

The same experiment with frwiki produced no noticable impact on
jobrunners. If it goes well again will try it on ewiki, and
eventually for all wikis.

Once we know if the jobrunners can handle the parsing load, the
wgTemporaryParsoidHandlerParserCacheWriteRatio should be removed
entirely. This will re-enable cache writes.

Bug: T339867
Change-Id: I2d44c711975cf151910940a5393b9b0668b11625