CommonsSettings.php: Use $wgCopyUploadProxy, not $wgLocalHTTPProxy
Why:
In 1534d16af8, we switched the ExtensionDistributor extension to use the
proxy in order to access gerrit in a way that remains transparent to the
extension and yet allows to move to mw-on-k8s without causing undue
burden to the owners of Gerrit/Gitlab. The full reasoning is in:

https://phabricator.wikimedia.org/T340483#8966576

However, wgLocalHTTPProxy was the wrong setting. That variable points
back to MediaWiki itself in the WMF infrastructure. It was introduced in
T288848 to allow for that function to work.

What:
Switch to using wgCopyUploadProxy

Change-Id: I167c1051c091cd0b65d2157396b53613c6c09e44