Disable PC writes for parsoid endpoints
The parsoid endpoints are primarily hit by restbase when pre-generating
page content. Having them write to the parser cache is redundant to
having ParsoidCachePrewarmJob do the same. If the parsoid endpoints no
longer do it, the jobrunners will have to do the parsing work.

This patch is an experiment to see the impact this would have on the
load on the jobrunner cluster.

When we tested this with frwiki, dewiki, and eventually enwiki, there
was no noticable impact.

Once we know that the jobrunners can handle the parsing load, we
should re-enable cache writes.

Bug: T339867
Change-Id: Id196adbfe5a0a66784836f54767420c87343f5be