Set $wgWBRepoSettings['defaultEntityNamespaces'] to false
This setting controls whether Wikibase will create Item: and Property:
namespaces on its own. In production, we don’t want this (we use the
main namespace for Items, and create the Property: namespace ourselves),
but it’s useful for other Wikibase installs. Currently, Wikibase has
this setting false by default (i.e., this change is currently a no-op),
and sets it to true in the example settings; since we want to get rid of
the need for example settings, we want to change the default to true,
but for that we first need to set it to false explicitly in production.

Done without a wmg global in InitialiseSettings.php – I don’t think
that’s needed here. (We unconditionally set custom entitySources, i.e.
don’t have an isset( $wmgWikibaseEntitySources ) check there, so we
should also unconditionally deactivate the default entity sources via
the defaultEntityNamespaces setting.)

Bug: T291617
Change-Id: I9e79b46e9f2fc3b764a074c11ae9034d321f1b53