Remove $wgNamespacesWithSubpages overrides on the MediaWiki namespace
History:

- Before MediaWiki 1.16, the MediaWiki namespace defaults to not having
  subpages. https://www.mediawiki.org/wiki/Manual:$wgNamespacesWithSubpages#Default_value
- Initially, InitialiseSettings.php specified a complete list for all
  wikis, at that time, the default value `8 => 0` repeated for most of
  wikis, the value didn't get updated when the default changed in 1.16.
- Commit d17b7789da802f57e7cdad9ea7a900af063ef514 cleaned up some
  overrides and 97feb2e65be4ce9e7668ec3643c91d0990c4a52d stated the
  default value (true) and removed `8 => 1` for some wikis which
  explicitly set before.

We should remove the outdated overrides to reflect the reality that
for a message, the pages prefixed with its name and slash are its
subpage.

Keeping the subpages on the MediaWiki namespace disabled caused
inconvenience when we want to check a message's overrides in different
languages, as the subpage count and the link to Special:PrefixIndex
in the page info are missing.

Bug: T340697
Change-Id: I10454a13033a59de2c009069e03376bd0c3f097e