PageView: Route requests through restbase service proxy
k8s containers are unable to reach the wikimedia.org
URL directly, and we need to be using the appropriate
service proxy instead. Do that by changing
PageViewInfoWikimediaEndpoint in PageView's configuration.

Keep the previous configuration for Beta, as the intention
seems to be for Beta to display production's pageview numbers.

Bug: T341191
Change-Id: If8c82f140b4d6a04f3803c3442f1b80867e8023c