mc: Remove mcrouter-with-onhost-tier from ParserCache
Switch ParserCache back to mysql + (regular) mcrouter, without a third
"mcrouter-with-onhost-tier" tier.

The third tier was introduced in 2020 as part of I514f0b881e9e5aa1
to reduce network congestion on network links between appservers
and memcached (T244340) and to avoid the associated outages those
could cause.

SRE and Perf Teams have since learned that this is not very effective
in practice. The original issue of network congestion has been
addressed since then with bigger links (10G), and the associated
risk of outages has additionally been reduced by also having memcached
"gutter pools" that can take over even if the links could still be
congested (or be unreachable for other reasons).

As such, the onhost-tier was left out from the MW-on-K8s deployment
(the onhost-tier no-ops there, alias to regular mcrouter).

To align with MW-on-K8s, disable use of onhost-tier on baremetal
appservers as well (in prod and Beta), and remove its cache ID.

After this, the mcrouter configuration in Puppet for onhost-tier,
and the memcached installations on appservers, can be removed as well.

Bug: T264604
Change-Id: I92dd062dc2c94a0ebdb2b2988d655e04095e36e8